+++ app/app.ts
@@ -1,9 +1,6 @@
-import dotenv from "dotenv"
 import express from "express"
 import { Application, Request, Response } from "express"
 import api from "../routes/api";
-
-dotenv.config()
 
 export const app: Application = express()