+++ app/controllers/ConverterController.ts
@@ -1,10 +1,27 @@
-import { Request, Response } from "express";
+import { Request, Response } from 'express';
+import { validate, convert } from '../helpers/converterHelper'
 
 export default {
-    convert: (req: Request, res: Response) => {
+    convert: async (req: Request, res: Response) => {
 
-        res.send({
-            result: {pe:3,pa:3}
-        })
+        let status: number = 200
+        const response: {[index: string]:any } = {}
+        const params: { from: string, to: string, amount: number, date?: string } = {
+            from: req.params.from,
+            to: req.params.to,
+            amount: parseInt(req.params.amount, 10),
+            date: req.params.date
+        }
+
+        const error: string = validate(params)
+        if (error) {
+            status = 400
+            response.message = error
+        } else {
+            response[req.params.from] = parseInt(req.params.amount, 10)
+            response[req.params.to] = await convert(params)
+        }
+
+        res.send(response).status(status)
     }
 }