+++ routes/api.ts
@@ -3,6 +3,6 @@ import ConverterController from '../app/controllers/ConverterController'
 const router = Router()
 
 // Api routes
-router.get('/convert', ConverterController.convert)
+router.get('/convert/:from/:to/:amount/:date?', ConverterController.convert)
 
 export default router;