+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for Converter controller added
 - Code coverage added
 - Typescript lint validation added
+- Simple convert feature added